+++ src/todomato.js
@@ -1,12 +1,14 @@
 /** @format */
 
+import { addTodoToDOM } from './add_tido_to_dom/addTodoToDOM.js';
 import { ControllerTimer } from './controllers/Timer.js';
 import { ControllerTodo } from './controllers/Todo.js';
+import { loadTodos, loadCompleted } from './load_todo_list/loatTodoList.js';
 
 const todoForm = document.getElementById('todo-form');
 const todoNameInput = document.getElementById('todo-name');
 const todoDescriptionInput = document.getElementById('todo-description');
-const todoList = document.getElementById('todo-list');
+
 const logoutButton = document.getElementById('btnLogOut');
 
 logoutButton.addEventListener('click', () => {
@@ -34,31 +36,30 @@ document.addEventListener('DOMContentLoaded', () => {
 const controllerTodo = new ControllerTodo();
 const controllerTimer = new ControllerTimer();
 
-document.addEventListener('DOMContentLoaded', () => {
-  const completedList = document.getElementById('completed-list');
-  completedList.innerHTML = '';
+// document.addEventListener('DOMContentLoaded', () => {
+//   const completedList = document.getElementById('completed-list');
+//   completedList.innerHTML = '';
 
-  let currentTodos = localStorage.getItem('todosCompleted');
-  let currentUser = localStorage.getItem('userLogged');
+//   let currentTodos = localStorage.getItem('todosCompleted');
+//   let currentUser = localStorage.getItem('userLogged');
 
-  if (currentUser && currentTodos) {
-    let objectUser = JSON.parse(currentUser);
-    let todosCompleted = JSON.parse(currentTodos);
+//   if (currentUser && currentTodos) {
+//     let objectUser = JSON.parse(currentUser);
+//     let todosCompleted = JSON.parse(currentTodos);
 
-    let userId = objectUser.id;
+//     let userId = objectUser.id;
 
-    console.log('User ID:', userId);
+//     console.log('User ID:', userId);
 
-    let userTodosCompleted = todosCompleted.filter(todo => todo.userId === userId);
+//     let userTodosCompleted = todosCompleted.filter(todo => todo.userId === userId);
 
-    console.log("Todos completati per l'utente:", userTodosCompleted);
+//     console.log("Todos completati per l'utente:", userTodosCompleted);
 
-    const controllerTodo = new ControllerTodo();
-    controllerTodo.loadTodoCompleted(userTodosCompleted);
+//     loadCompleted(userTodosCompleted);
 
-    console.log('Caricamento todos completati riuscito!');
-  }
-});
+//     console.log('Caricamento todos completati riuscito!');
+//   }
+// });
 
 todoForm.addEventListener('submit', function (event) {
   event.preventDefault();
@@ -68,7 +69,7 @@ todoForm.addEventListener('submit', function (event) {
   let userId = currentUser.id;
 
   const todo = controllerTodo.addTodo(nomeAttivita, description, userId);
-  addTodoToDOM(todo);
+  addTodoToDOM(todo, controllerTodo, controllerTimer);
 
   console.log(todo);
   console.log('Todo utente Salvato');
@@ -77,182 +78,10 @@ todoForm.addEventListener('submit', function (event) {
   todoDescriptionInput.value = '';
 });
 
-function addTodoToDOM(todo) {
-  const li = document.createElement('li');
-  li.setAttribute('data-id', todo.todoId);
-
-  const nameSpan = document.createElement('span');
-  nameSpan.className = 'todo-name';
-  nameSpan.textContent = todo.nomeAttivita;
-
-  const descriptionSpan = document.createElement('span');
-  descriptionSpan.className = 'todo-description';
-  descriptionSpan.textContent = `: ${todo.description}`;
-
-  const timerSpan = document.createElement('span');
-  timerSpan.className = 'timer';
-  timerSpan.textContent = '(--:--)';
-
-  const timerControls = document.createElement('div');
-  timerControls.className = 'timer-controls';
-
-  const resumeButton = document.createElement('button');
-  const pauseButton = document.createElement('button');
-  const editButton = document.createElement('button');
-  const deleteButton = document.createElement('button');
-  const completeButton = document.createElement('button');
-  const createTimerButton = document.createElement('button');
-
-  resumeButton.style.display = 'none';
-  const resumeImg = document.createElement('img');
-  resumeImg.src = './assets/img/clock-02.svg';
-  resumeImg.width = 20;
-  resumeImg.height = 20;
-  resumeImg.alt = 'Resume';
-  resumeButton.appendChild(resumeImg);
-  resumeButton.addEventListener('click', function () {
-    resumeButton.style.display = 'none';
-    pauseButton.style.display = 'block';
-    if (todo.timer) {
-      controllerTimer.resume(todo.timer);
-    }
-  });
-
-  pauseButton.style.display = 'none';
-  const pauseImg = document.createElement('img');
-  pauseImg.src = './assets/img/pause-02.svg';
-  pauseImg.width = 20;
-  pauseImg.height = 20;
-  pauseImg.alt = 'Pause';
-  pauseButton.appendChild(pauseImg);
-  pauseButton.addEventListener('click', function () {
-    resumeButton.style.display = 'block';
-    pauseButton.style.display = 'none';
-    if (todo.timer) {
-      controllerTimer.pause(todo.timer);
-    }
-  });
-
-  editButton.style.display = 'none';
-  const editImg = document.createElement('img');
-  editImg.src = './assets/img/editButton-02.svg';
-  editImg.width = 20;
-  editImg.height = 20;
-  editImg.alt = 'Edit';
-  editButton.appendChild(editImg);
-  editButton.addEventListener('click', function () {
-    controllerTimer.pause(todo.timer);
-    const newName = prompt('Inserisci un nuovo nome attività:', todo.nomeAttivita);
-    const newDescription = prompt('Inserisci una nuova descrizione:', todo.description);
-
-    if (newName !== null) {
-      controllerTodo.updateTodoName(todo.todoId, newName);
-      nameSpan.textContent = newName;
-    }
-
-    if (newDescription !== null) {
-      controllerTodo.updateTodoDescription(todo.todoId, newDescription);
-      descriptionSpan.textContent = `: ${newDescription}`;
-    }
-    controllerTimer.resume(todo.timer);
-  });
-
-  deleteButton.style.display = 'none';
-  const deleteImg = document.createElement('img');
-  deleteImg.src = './assets/img/eraseButton-03.svg';
-  deleteImg.width = 20;
-  deleteImg.height = 20;
-  deleteImg.alt = 'Delete';
-  deleteButton.appendChild(deleteImg);
-  deleteButton.addEventListener('click', function () {
-    controllerTodo.deleteTodo(todo.todoId);
-    li.remove();
-  });
-
-  completeButton.style.display = 'none';
-  const completeImg = document.createElement('img');
-  completeImg.src = './assets/img/checkCompleted-02.svg';
-  completeImg.width = 20;
-  completeImg.height = 20;
-  completeImg.alt = 'Complete';
-  completeButton.appendChild(completeImg);
-  completeButton.addEventListener('click', function () {
-    controllerTodo.moveTodoToCompleted(todo);
-    li.remove();
-  });
-
-  const playImg = document.createElement('img');
-  playImg.src = './assets/img/play-02.svg';
-  playImg.width = 20;
-  playImg.height = 20;
-  playImg.alt = 'Create Timer';
-  createTimerButton.appendChild(playImg);
-  createTimerButton.addEventListener('click', function () {
-    pauseButton.style.display = 'block';
-    deleteButton.style.display = 'block';
-    editButton.style.display = 'block';
-    completeButton.style.display = 'block';
-
-    createTimerButton.style.display = 'none';
-    if (!todo.timer) {
-      const timer = controllerTimer.create(25, true);
-      todo.timer = timer;
-    }
-  });
-
-  timerControls.appendChild(createTimerButton);
-  timerControls.appendChild(pauseButton);
-  timerControls.appendChild(resumeButton);
-  timerControls.appendChild(deleteButton);
-  timerControls.appendChild(editButton);
-  timerControls.appendChild(completeButton);
-
-  li.appendChild(nameSpan);
-  li.appendChild(descriptionSpan);
-  li.appendChild(timerSpan);
-  li.appendChild(timerControls);
-  todoList.appendChild(li);
-
-  const intervalId = setInterval(() => {
-    if (todo.timer && !todo.timer.isPaused) {
-      const remainingTime = controllerTimer.read(todo.timer);
-      if (remainingTime.minutes > 0 || remainingTime.seconds > 0) {
-        if (remainingTime.seconds < 10) {
-          timerSpan.textContent = ` (${remainingTime.minutes} : 0${remainingTime.seconds})`;
-        } else {
-          timerSpan.textContent = ` (${remainingTime.minutes} : ${remainingTime.seconds})`;
-        }
-      } else {
-        if (todo.timer.isPomodoro) {
-          todo.timer = controllerTimer.nextPomodoroSession(todo.timer);
-          const newTimeRemaining = controllerTimer.read(todo.timer);
-          timerSpan.textContent = ` (${newTimeRemaining.minutes} : ${newTimeRemaining.seconds})`;
-        } else {
-          timerSpan.textContent = " (Time's up!)";
-          controllerTodo.moveTodoToCompleted(todo);
-          clearInterval(intervalId);
-        }
-      }
-    }
-  }, 1000);
-}
-
-function loadTodos() {
-  const todos = JSON.parse(localStorage.getItem('todos')) || [];
-
-  const addedTodoIds = new Set();
-
-  const todoList = document.getElementById('todo-list');
-
-  todos.forEach(todo => {
-    if (!addedTodoIds.has(todo.id)) {
-      addTodoToDOM(todo);
-      addedTodoIds.add(todo.id);
-    }
-  });
-}
-
-window.onload = loadTodos;
+window.onload = function () {
+  loadTodos(controllerTodo, controllerTimer, addTodoToDOM);
+  loadCompleted(controllerTodo);
+};
 
 const videoElement = document.getElementById('video-tag');
 const unmuteButton = document.getElementById('unmute-button');
